% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_uiddb.R
\name{build_uiddb}
\alias{build_uiddb}
\title{Build UID Database}
\usage{
build_uiddb(
  assemblies,
  dbpath = NULL,
  update_after = 30,
  verbose = getOption("verbose")
)
}
\arguments{
\item{assemblies}{character; vector of assembly accessions.}

\item{dbpath}{character; path to an exported uiddb database. \code{NULL} by
default in which case the database is not exported but remains in memory.}

\item{update_after}{numeric; if and entry in the database is older than this
value (in days) and the entry is queried again, then it will be updated in
the database.}

\item{verbose}{logical; print verbose messages to console}
}
\value{
A data frame of unique ID-s. If \code{dbpath} is not \code{NULL}
then the data frame is also exported as a .rda file and is read within
subsequent function calls.
}
\description{
This function connects GCA/GCF type assembly accessions with internal ID-s
within a number of NCBI databases.
}
\examples{
\dontrun{
ids <- c("GCA_003012895.2", "GCF_000695855.3")
uiddb <- build_uiddb(ids)
}
}
