% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_genome.R
\name{download_genome}
\alias{download_genome}
\title{Download Genomes from NCBI}
\usage{
download_genome(
  assembly_uid,
  type = "genomic.gbff",
  dirpath = NULL,
  from = "refseq",
  verbose = getOption("verbose")
)
}
\arguments{
\item{assembly_uid}{numeric; a unique identifier. Use \code{get_uid} to
acquire a unique UID.}

\item{type}{character; the file extension to download. Valid options are
\code{"assembly_report"}, \code{"assembly_stats"}, \code{"cds"},
\code{"feature_count"}, \code{"feature_table"}, \code{"genomic.fna"},
\code{"genomic.gbff"}, \code{"genomic.gff"}, \code{"genomic.gtf"},
\code{"protein.faa"}, \code{"protein.gpff"}, \code{"translated_cds"}.}

\item{dirpath}{character; the path to the directory where the file should be
downloaded.}

\item{from}{character; either \code{"refseq"} or \code{"genbank"}.
If \code{from} is set to \code{"refseq"} then the function will attempt to
download the file from the refseq repository and use \code{"genbank"} as a
fallback option.}

\item{verbose}{boolean; should verbose messages be printed to console?}
}
\description{
This function directly accesses a genome through the NCBI FTP server.
}
\note{
ToDo: vectorise and add input validation to ensure function fails early
if there are input errors. Add tests.
}
\examples{
\dontrun{
# Aim: download genbank file for "GCF_003007635.1"

# The function will access files within this directory:
# ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/003/007/635/

# Step 1 - Get NCBI Assembly UID
uid <- get_uid("GCF_003007635.1", db = "assembly")

# Step 2 - Download file
download_genome(uid, type = "genomic.gbff")
}
}
