% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_report.R
\name{download_report}
\alias{download_report}
\title{Download assembly reports}
\usage{
download_report(
  urls,
  report_dir = NULL,
  update_threshold = 30,
  verbose = getOption("verbose")
)
}
\arguments{
\item{urls}{character; a vector of assembly report urls.}

\item{report_dir}{character; download directory for the assembly reports.}

\item{update_threshold}{numeric; reports accessed earlier than this value
(days) will be downloaded again and the existing versions will be
overwritten. Set below \code{0} to force download all reports.}

\item{verbose}{logical; print messages to screen?}
}
\description{
This function takes a vector of assembly report urls, checks for their
presence in the assembly report directory and downloads them if they are
either missing from the directory or might be outdated.
}
\note{
This is the third step within the pipeline for downloading GenBank
files.
}
\examples{
\dontrun{
phages <- get_genomeid("Autographiviridae", db = "assembly")
report_url <- get_report_url(phages$ids[1], cache = FALSE, verbose = TRUE)
download_report(report_url, verbose = TRUE)
}
}
\seealso{
\code{get_genomeid},
\code{get_report_url()},
\code{parse_report()},
\code{extract_accn()},
\code{download_gb()}
}
