% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{dv_connect}
\alias{dv_connect}
\title{Connect to a Dataset on a Dataverse server}
\usage{
dv_connect(dataset, version = ":latest", server, apikey = "")
}
\arguments{
\item{dataset}{character, "persistentId" of the Dataset to connect to.}

\item{version}{character, version of the Dataset to connect to.}

\item{server}{character, URL of the Dataverse server.}

\item{apikey}{character, your API key for the Dataverse server.}

\item{...}{Additional arguments to pass to `dataverse::get_dataset()`.}
}
\value{
A dataset object object of class `dataverse_dataset`.
}
\description{
Connect to a Dataset on a Dataverse server
}
\details{
The persistentId and the server address can be obtained from the URL of the 
Dataset. The URL starts with the server address, and the persistentId is the 
value of the "persistentId" query parameter. For example, in the URL
"https://dataverse.no/dataset.xhtml?persistentId=doi:10.18710/NCKZD7", the 
server address is "https://dataverse.no" and the persistentId is 
"doi:10.18710/NCKZD7". If the Dataset is private, you will need to provide 
your API key to access it. You can obtain an API key by creating an account
on the Dataverse server and generating a new API token in your account 
settings.
}
\examples{
# connect to a public dataset on dataverse.no
conn <- dv_connect(
  dataset = "doi:10.18710/NCKZD7",
  server = "https://dataverse.no"
)
# connect to a private dataset on dataverse.no
conn <- dv_connect(
 dataset = "private_persistentId_here",
 server = "https://dataverse.no",
 apikey = "your_api_key_here"
)
}
