% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{dv_download}
\alias{dv_download}
\title{Download files from a Dataset of a Dataverse server}
\usage{
dv_download(files, dirpath = NULL, verbose = getOption("verbose"))
}
\arguments{
\item{files}{a `dv_files` object from [dv_list_files()].}

\item{dirpath}{character, path to the directory where files will be saved.
Defaults to the current working directory.}

\item{verbose}{logical, should verbose messages be printed to the console?}
}
\value{
A character vector of file paths for successfully downloaded files,
or `NA` for files that failed to download. Returned invisibly.
}
\description{
Download files from a Dataset of a Dataverse server
}
\examples{
conn <- dv_connect_dataset(
  dataset = "doi:10.18710/NCKZD7",
  server  = "https://dataverse.no"
)
conn |> dv_list_files(extension = "fasta") |> dv_download()
}
\seealso{
[dv_connect_dataset()], [dv_list_files()]
}
