% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{dv_list_files}
\alias{dv_list_files}
\title{List files in a Dataset on a Dataverse server}
\usage{
dv_list_files(conn)
}
\arguments{
\item{conn}{a `dv_connection` object from [dv_connect()].}
}
\value{
A tibble with information about the files in the Dataset.
}
\description{
List files in a Dataset on a Dataverse server
}
\examples{
conn <- dv_connect(
  dataset = "doi:10.18710/NCKZD7",
  server  = "https://dataverse.no"
)
conn |> dv_list_files()
}
\seealso{
[dv_connect()], [dv_download()]
}
