% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ena_download_reads.R
\name{ena_download_reads}
\alias{ena_download_reads}
\title{Download sequencing reads from the European Nucleotide Archive (ENA)}
\usage{
ena_download_reads(
  accession,
  type = "fastq",
  dirpath = NULL,
  mirror = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{accession}{character; a character vector of ENA Accession IDs.}

\item{type}{character; A character string specifying the type of file to
download, either \code{"run"}, \code{"fastq"} or \code{"err"}. See Details
for more  information.}

\item{dirpath}{character; the path to the directory where the file should be
downloaded. If \code{NULL}, download file to the working directory.}

\item{mirror}{logical; should the download directory mirror the structure of 
the FTP directory?}

\item{verbose}{logical; should verbose messages be printed to console?}
}
\description{
Download sequencing reads from the European Nucleotide Archive (ENA)
}
\details{
If \code{type == "run"} the function will download the read files that were 
originally submitted to ENA. If a run was originally submitted to another 
INSDC database (NCBI SRA, DDBJ) then a file of this category will not be 
available. If \code{type == "fastq"} the function will download one or more 
FASTQ files for each run. If \code{type = "err"} the function will download
files that can be used with NCBI's SRA Toolkit.
}
\examples{
\dontrun{
ena_download_reads("ERR1649607", type = "fastq", verbose = TRUE)
}

}
\references{
https://ena-docs.readthedocs.io/en/latest/retrieval/file-download/sra-ftp-structure.html
}
