% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ena_query.R
\name{ena_query}
\alias{ena_query}
\title{Retrieve sequences from ENA}
\usage{
ena_query(
  accessions,
  mode = "fasta",
  expanded = FALSE,
  annotation_only = FALSE,
  line_limit = 0,
  download = FALSE,
  destfile_by = "all",
  gzip = FALSE,
  set = FALSE,
  range = NULL,
  complement = FALSE,
  batch_size = 0,
  verbose = getOption("verbose")
)
}
\arguments{
\item{accessions}{character; Accessions to query.}

\item{mode}{character; Can be either \code{"embl"}, \code{"fasta"}, or
\code{"xml"}.}

\item{expanded}{logical; Get expanded records for CON sequences.}

\item{annotation_only}{logical; Only retrieve annotation, no sequence.}

\item{line_limit}{integer; Limit the number of text lines returned.}

\item{download}{logical; Download the result as a file.}

\item{destfile_by}{character; Number of files to download. 
\code{"batch"}: one file for each batch, \code{"all"}: one file altogether.}

\item{gzip}{logical; Download the result as a gzip file.}

\item{set}{logical; ???}

\item{range}{character; ???}

\item{complement}{logical; ???}

\item{batch_size}{integer; Number of accessions to query in a single request.
Using this value, accessions will be broken down into one or more batches.
If set to 0, all accessions will be queried in a single request.}

\item{verbose}{logical; Should verbose messages be printed to the console?}
}
\description{
Retrieve sequences from ENA
}
\examples{
\dontrun {
ena_query("LC136852")
ena_query(c("LC136852", "LC136853"))
}
}
