% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{examples}
\alias{examples}
\title{Examples}
\format{
A list with 2 elements:
\describe{
  \item{assembly}{NCBI Assembly}
  \item{biosample}{NCBI BioSample}
}
}
\usage{
examples
}
\description{
This data set contains a list of IDs which can be used to access data from
various data sources. These IDs are used across the package in function
documentations, tests, vignettes.
}
\keyword{datasets}
