% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_files.R
\name{flag_files}
\alias{flag_files}
\title{Flag files to keep for analysis}
\usage{
flag_files(filenames)
}
\arguments{
\item{filanames}{character; a character vector of filenames. Currently the
function only supports GCA/GCF identifiers. Look at the examples for more
details.}
}
\value{
The function returns a data frame where each file is listed in the
first column and the recommendation to keep the file for further analysis
is listed in the last column.
}
\description{
Some functions may download files that only differ in their source (e.g. GCA
from GenBank assemblies or GCF for RefSeq assemblies) or their version number
(v1, v2, etc.). This function helps remove redundant files by flagging which
files should be kept for further analysis.
}
\details{
The function first prioritises GCF over GCA and then the highest
version number.
}
\examples{
# keep GCF
filenames <- c("GCA_003012895.2_ASM301289v2_genomic.fna",
               "GCF_003012895.2_ASM301289v2_genomic.fna")
flag_files(filenames)                
               
# keep GCF even when version number is lower
filenames <- c("GCA_003012895.2_ASM301289v2_genomic.fna",
               "GCF_003012895.1_ASM301289v1_genomic.fna")
flag_files(filenames) 
               
filenames <- c("GCA_003012895.1_ASM301289v1_genomic.fna",
               "GCA_003012895.2_ASM301289v2_genomic.fna") 
flag_files(filenames) 
}
