% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{Bind rows in a list of list of data frames}
\usage{
flatten(x)
}
\arguments{
\item{x}{list; a list which contains a named list of data frames}
}
\value{
a flat list of data frames where lists of data frames with the same 
name are merged.
}
\description{
This is a utility function that converts a list of lists, each containing
a names list of data frames into a single flat list of data frames.
}
