% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Get metadata for biological sequences}
\usage{
get_meta(uid, db = "assembly")
}
\arguments{
\item{uid}{character; the internal UID of an entry inside a database.}

\item{db}{character; the database to search in. Can be either 
\code{"assembly"}, \code{"biosample"} or \code{"sra"}.}
}
\value{
a data frame
}
\description{
This function accesses metadata that usually accompanies the biological
sequences. Depending on the database queried this metadata can contain a
number of identifiers of information about the sample e.g. when and where it
was collected. Currently the function works with a few NCBI databases.
}
\examples{
\dontrun{
# search for the species in the NCBI Assembly database
assembly_uid <- get_uid("Dechloromonas phosphoritropha", db = "assembly")
# get assembly metadata for the first two hits
assembly_meta <- get_meta(assembly_uid$uid[1:2], db = "assembly")
# convert UIDs in the Assembly database to UIDs in the Biosample database
biosample_uid <- link_uid(assembly_uid$uid, from = "assembly", to = "biosample")
# get biosample metadata for the first two hits
biosample_meta <- get_meta(biosample_uid$result[1:2], db = "biosample")
}
}
