% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_url.R
\name{get_report_url}
\alias{get_report_url}
\title{Get the URL of an assembly report}
\usage{
get_report_url(
  uid,
  batch_size = 10,
  cache = FALSE,
  cache_file = NULL,
  update_threshold = 30,
  verbose = getOption("verbose")
)
}
\arguments{
\item{uid}{numeric; a vector of internal IDs.}

\item{batch_size}{the number of IDs to query in one batch. The function will
stop with an error if \code{batch_size} is too high. Decrease if necessary.}

\item{cache}{logical; should results be exported to hard drive?}

\item{cache_file}{character; file path for the exported file.}

\item{update_threshold}{numeric; entries accessed earlier than this value
(days) will be updated in the cached database. Set below \code{0} to force
the update.}

\item{verbose}{logical; print messages to screen?}
}
\description{
This function can be used to fetch urls that point to assembly reports.
}
\note{
This is the second step within the pipeline for downloading GenBank
files.

Currently the function only works with ID's that refer to entries
within the NCBI Assembly database.
}
\examples{
\dontrun{
phages <- get_genomeid("Autographiviridae", db = "assembly")
report_url <- get_report_url(phages$ids[1], cache = FALSE, verbose = TRUE)
}
}
\seealso{
\code{get_genomeid},
\code{download_report()},
\code{parse_report()},
\code{extract_accn()},
\code{download_gb()}
}
