% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uid.R
\name{get_uid}
\alias{get_uid}
\title{Collect internal IDs from NCBI databases}
\usage{
get_uid(term, db = "assembly")
}
\arguments{
\item{term}{character; the search term. Check details for more information.
\code{term} takes a vector of strings and collapses them with " OR ".}

\item{db}{character; the database to search in. For options see
\code{rentrez::entrez_dbs()}}
}
\value{
an object of class \code{esearch}.
}
\description{
This function replicates the NCBI website's search utility. The search term
is matched against the chosen database and the function returns a list of
internal IDs that can be used e.g. to link NCBI entries with entries in other
NCBI databases.
}
\note{
This is the first step within the pipeline for downloading GenBank
files.
}
\examples{
\dontrun{
get_uid(c("Autographiviridae", "Podoviridae"), db = "assembly")
}
}
\seealso{
\code{get_report_url()},
\code{download_report()},
\code{parse_report()},
\code{extract_accn()},
\code{download_gb()}
}
