% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_uids.R
\name{link_uids}
\alias{link_uids}
\title{Convert between NCBI UIDs}
\usage{
link_uids(uids, from, to)
}
\arguments{
\item{uids}{character; a vector of uids.}

\item{from}{character; the database the queried uids come from.
\code{rentrez::entrez_dbs()} lists all available options.}

\item{to}{character; the database in which the function should look for links.
\code{rentrez::entrez_dbs()} lists all available options.}
}
\value{
A character vector of linked uids from the new database.
}
\description{
Each entry in an NCBI database has its unique internal id. Entries in
different databases may be linked. For example, an assembly in the NCBI
Assembly database may be linked with metadata in the NCBI BioSample database.
This function links uids from one database with uids from another.
}
\examples{
\dontrun{
link_uids("4253631", "assembly", "biosample")
}
}
