% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgnify_endpoints.R
\name{mgnify_endpoints}
\alias{mgnify_endpoints}
\title{MGnify endpoints}
\usage{
mgnify_endpoints()
}
\value{
a tibble of API-s and their respective endpoints
}
\description{
This functions queries MGnify for all available endpoints
}
\note{
The function prints contents of the following url:
\url{https://www.ebi.ac.uk/metagenomics/api/v1/}
}
\examples{
\dontrun{
mgnify_endpoints()
}
}
