% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgnify_list.R
\name{mgnify_list}
\alias{mgnify_list}
\title{Retrieve a list of instances from MGnify}
\usage{
mgnify_list(
  query,
  from,
  from_id,
  page = NULL,
  sleep = 0.2,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{character; what to look for.}

\item{from}{character; API. See \code{"mgnify_endpoints()"}.}

\item{from_id}{character; more precise filtering for the API.}

\item{page}{numeric; the API's response is paginated this tells the API which
page to return. If \code{NULL}, the function will return all pages.}

\item{sleep}{character; number of seconds to sleep before requesting the next
page.}

\item{verbose}{logical; should verbose messages be printed to console?}
}
\description{
This function retrieves a list of identifiers to look up with other
functions.
}
\examples{
\dontrun{
# Query samples collected from biogas plants
mgnify_list(query = "samples",
            from = "biomes",
            from_id = "root:Engineered:Biogas plant",
            page = 1)
}
}
