% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi2ena.R
\name{ncbi2ena}
\alias{ncbi2ena}
\title{Convert accessions from NCBI to ENA}
\usage{
ncbi2ena(accessions, type)
}
\arguments{
\item{accessions}{character; a vector or ENA accessions.}

\item{type}{character; type of accessions. Supported types: `biosample`, 
`bioproject`.}
}
\value{
A tibble with two columns, `ncbi` and `ena`.
}
\description{
Take a vector of NCBI accessions and convert them to ENA accessions.
}
\examples{
ncbi2ena("SAMEA111452506", type = "biosample")
}
