% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_download_genome.R
\name{ncbi_download_genome}
\alias{ncbi_download_genome}
\title{Download Genomes from NCBI Assembly Database}
\usage{
ncbi_download_genome(
  query,
  type = "genomic.fna",
  dirpath = NULL,
  mirror = FALSE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{an object of class `ncbi_uid`, `ncbi_uid_link`, `ncbi_link`, or 
an integer vector of NCBI Assembly UIDs. See Details for more information.}

\item{type}{character; the file extension to download. Valid options are
\code{"assembly_report"}, \code{"assembly_stats"}, \code{"cds"},
\code{"feature_count"}, \code{"feature_table"}, \code{"genomic.fna"},
\code{"genomic.gbff"}, \code{"genomic.gff"}, \code{"genomic.gtf"},
\code{"protein.faa"}, \code{"protein.gpff"}, \code{"translated_cds"}.}

\item{dirpath}{character; the path to the directory where the file should be
downloaded. If \code{NULL}, download file to the working directory.}

\item{mirror}{logical; should the download directory mirror the structure of 
the FTP directory?}

\item{verbose}{logical; should verbose messages be printed to console?}
}
\description{
This function directly downloads genome data through the NCBI FTP server.
}
\details{
`ncbi_get_uid()` returns an object of class `ncbi_uid`; 
`ncbi_link_uid` returns an object of class `ncbi_uid_link`; `ncbi_link`
returns and object of class `ncbi_link`. These objects may be used directly 
as query input for `ncbi_download_genome`. It is recommended to use this
approach. Alternatively, you can also use a character vector of UIDs as query
input.  This approach is not recommended because there are no consistency 
checks, the function will just attempt to interpret the query as NCBI 
Assembly UIDs.
}
\examples{
\dontrun{
# Download a single genome
ncbi_get_uid("GCF_003007635.1", db = "assembly") |>
  ncbi_download_genome()

"SAMN08619567" |>
  ncbi_get_uid(db = "biosample") |>
  ncbi_link_uid(to = "assembly") |>
  ncbi_download_genome()
  
"SAMN08619567" |>
  ncbi_link(from = "biosample", to = "assembly") |>
  ncbi_download_genome()

# Download multiple genomes, mirror FTP directory structure
data(examples) 

examples$assembly |> 
  ncbi_get_uid(db = "assembly") |>
  ncbi_download_genome()
}
}
