% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_get_meta.R
\name{ncbi_get_meta}
\alias{ncbi_get_meta}
\title{Get sequence metadata from NCBI}
\usage{
ncbi_get_meta(
  query,
  db = NULL,
  batch_size = 100,
  use_history = TRUE,
  parse = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{either an object of class \code{ncbi_uid} or an integer vector 
of UIDs. See Details for more information.}

\item{db}{character; the database to search in. For options see
\code{ncbi_dbs()}. Note that not all of these databases are supported.}

\item{batch_size}{integer; the number of search terms to query at once. If
the number of search terms is larger than \code{batch_size}, the search terms
are split into batches and queried separately.}

\item{use_history}{logical; should the function use web history for faster
API queries?}

\item{parse}{logical; should the function attempt to parse the output into a
tibble?}

\item{verbose}{logical; Should verbose messages be printed to console?}
}
\value{
If \code{parse = FALSE} the function will return an object of class
\code{ncbi_meta}, which is a character vector with some extra information
about the database. This output can be used directly with \code{ncbi_parse}.
If \code{parse = TRUE} the function will attempt to parse the data using 
\code{ncbi_parse}. If parsing is successful, the function will return a
tibble, otherwise it will return the unparsed \code{ncbi_meta} object.
}
\description{
This function retrieves sequence metadata from a given NCBI sequence
database.
}
\details{
You can give UIDs to \code{ncbi_get_meta()} in two ways: 1. You can
use functions like \code{ncbi_get_uid()} or \code{ncbi_link_uid} to get UIDs,
and then use the returned \code{ncbi_uid} objects directly with
\code{ncbi_get_meta}. If you follow this approach then you do not have to
specify the \code{db} argument since the function can extract it from the
\code{ncbi_uid} object. However, if you do provide it, then it must be
identical to the \code{db} attribute of the \code{"ncbi_uid"} object. 2. 
Alternatively, you can just provide a vector of UIDs, but then you must 
specify the \code{db} argument as well.
}
\examples{
\dontrun{
data(examples)
uids <- ncbi_get_uid(examples$biosample, db = "biosample")
meta <- ncbi_get_meta(uids)
}
}
