% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_get_summary.R
\name{ncbi_get_summary}
\alias{ncbi_get_summary}
\title{Get object summary from NCBI}
\usage{
ncbi_get_summary(
  query,
  db = NULL,
  batch_size = 100,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{either an object of class \code{ncbi_uid} or an integer vector 
of UIDs. See Details for more information.}

\item{db}{character; the database to search in. For options see
\code{ncbi_dbs()}.}

\item{batch_size}{integer; the number of items to query at once. If query 
length is larger than `batch_size`, the query will be split into batches.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
A list of rentrez summary objects.
}
\description{
This function retrieves object summary data from a given NCBI database.
}
\details{
If query is an `ncbi_uid` object, the `db` argument is optional. If 
`db` is not specified, the function will retrieve it from the query object. 
However, if it is specified, it must be identical to the `db` attribute of 
the query.
}
\examples{
\dontrun{
assemblies <- c("GCF_000002435.2", "GCF_000299415.1")
uids <- ncbi_get_uid(assemblies, db = "assembly")
ncbi_get_summary(uids)
}
}
