% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_get_uid.R
\name{ncbi_get_uid}
\alias{ncbi_get_uid}
\title{Collect UID-s from NCBI databases}
\usage{
ncbi_get_uid(
  term,
  db = "assembly",
  cache_file = NULL,
  verbose = getOption("verbose")
)
}
\arguments{
\item{term}{character; one or more search terms.}

\item{db}{character; the database to search in. For options see
\code{rentrez::entrez_dbs()}}

\item{cache_file}{character; the name of the cache file without the file
extension. If \code{NULL}, results are not cached.}

\item{verbose}{logical; should verbos messages be printed to the console?}
}
\value{
A tibble.
}
\description{
This function replicates the NCBI website's search utility. One or more
search terms are matched against the chosen database and the function returns
a tibble of internal NCBI UID-s that can be used e.g. to link NCBI entries
with entries in other NCBI databases.
}
\examples{
\dontrun{
ncbi_get_uid("GCA_003012895.2")
ncbi_get_uid("Autographiviridae OR Podoviridae", db = "assembly")
ncbi_get_uid(c("WP_093980916.1", "WP_181249115.1"), db = "protein")
}
}
