% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_link.R
\name{ncbi_link}
\alias{ncbi_link}
\title{Link ID-s from one NCBI database to another}
\usage{
ncbi_link(query, from, to, batch_size = 100, verbose = getOption("verbose"))
}
\arguments{
\item{query}{character; a vector of IDs}

\item{from}{character; the database the queried ID-s come from.
\code{ncbi_dbs()} lists all available options.}

\item{to}{character; the database in which the function should look for links.
\code{ncbi_dbs()} lists all available options.}

\item{batch_size}{integer; the number of search terms to query at once. If
the number of search terms is larger than \code{batch_size}, the search terms
are split into batches and queried separately.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
A tibble with two columns. The first column contains IDs in the 
`from` database, the second column contains linked IDs in the `to` database.
}
\description{
Each entry in an NCBI database has its unique ID. Entries in different 
databases may be linked. For example, entries in the NCBI Assembly database
may be linked with entries in the NCBI BioSample database. This function 
attempts to link ID-s from one database to another.
}
\examples{
\dontrun{
ncbi_link("GCF_000002435.2", from = "assembly", to = "biosample")
ncbi_link("SAMN02714232", from = "biosample", to = "assembly")
}
}
