% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_link_uids.R
\name{ncbi_link_uids}
\alias{ncbi_link_uids}
\title{Link NCBI UID-s}
\usage{
ncbi_link_uids(
  query,
  from,
  to,
  cache_file = NULL,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{character; a vector of uids.}

\item{from}{character; the database the queried uids come from.
\code{rentrez::entrez_dbs()} lists all available options.}

\item{to}{character; the database in which the function should look for links.}

\item{cache_file}{character; the name of the cache file without the file
extension. If \code{NULL}, results are not cached.}

\item{verbose}{logical; should verbos messages be printed to the console?
\code{rentrez::entrez_dbs()} lists all available options.}
}
\value{
A tibble
}
\description{
Each entry in an NCBI database has its unique internal id. Entries in
different databases may be linked. For example, an assembly in the NCBI
Assembly database may be linked with metadata in the NCBI BioSample database.
This function links uids from one database with uids from another.
}
\examples{
\dontrun{
ncbi_link_uids("4253631", "assembly", "biosample")
ncbi_link_uids(c("1226742659", "1883410844"), "protein", "nuccore")
}
}
