% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_meta.R
\name{ncbi_meta}
\alias{ncbi_meta}
\title{Collect NCBI BioSample metadata}
\usage{
ncbi_meta(biosample_uids, verbose = getOption("verbose"))
}
\arguments{
\item{biosample_uids}{character; a character vector of BioSample UID-s.}

\item{verbose}{logical; Should verbose messages be printed to console?}
}
\description{
This function collects metadata for entries in NCBI BioSample database.
}
\examples{
\dontrun{
# Sample metadata from BioSample ID
biosample_uid <- get_uid("SAMN02714232", db = "biosample")
ncbi_meta(biosample_uid)

# Sample metadata from assembly accession 
assembly_uid <- get_uid("GCF_000695855.3")
biosample_uid <- link_uids(assembly_uid, from = "assembly", to = "biosample")
ncbi_meta(biosample_uid)
}
}
