% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_meta_assembly.R
\name{ncbi_meta_assembly}
\alias{ncbi_meta_assembly}
\title{Retrieve NCBI Assembly metadata}
\usage{
ncbi_meta_assembly(assembly_uid)
}
\arguments{
\item{assembly_uid}{numeric}
}
\description{
Retrieve NCBI Assembly metadata
}
\examples{
\dontrun{
ncbi_meta_assembly(419738)
}
}
