% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_parse.R
\name{ncbi_parse}
\alias{ncbi_parse}
\title{Parse NCBI sequence metadata}
\usage{
ncbi_parse(meta, db, format = "xml", verbose = getOption("verbose"))
}
\arguments{
\item{meta}{character; either an unparsed metadata object returned by
\code{ncbi_get_meta()} or the path to a file that was downloaded from NCBI.}

\item{db}{character; the NCBI database from which the data was retrieved.}

\item{format}{character; the format of the data set. Currently only
\code{"xml"} is supported.}

\item{verbose}{logical; Should verbose messages be printed to console?}
}
\value{
a tibble.
}
\description{
This function can be used to parse various retrieved non-sequence data sets
from NCBI into a tibble. These data sets usually accompany the biological
sequences and contain additional information e.g. identifiers, information
about the sample, the sequencing platform, etc.
}
\details{
This function is integrated into \code{ncbi_get_meta()} and is 
called automatically if \code{parse = TRUE} (default). However, it can also
be used  separately e.g. when you want to examine the unparsed metadata
object before parsing, or when you already downloaded the metadata manually
and you just want to parse it into a tabular format.
}
\examples{
\dontrun{
data(examples)
#' 
# NCBI Assembly, download XML file from NCBI and parse

# Manually download the XML file
# https://www.ncbi.nlm.nih.gov/assembly/GCF_000299415.1
# upper right corner -> send to -> file -> format = xml -> create file
# Parse XML
ncbi_parse(meta = "assembly_summary.xml", db = "assembly", format = "xml")

# NCBI BioSample, fully programmatic access, separate retrieval and parsing

# Get metadata but do not parse
meta <- ncbi_get_meta(examples$biosample, db = "biosample", parse = FALSE)
# Parse metadata separately
ncbi_parse(meta = meta, db = "biosample", format = "xml")

# NCBI BioSample, download XML file from NCBI and parse

# Manually download the XML file
# https://www.ncbi.nlm.nih.gov/biosample/?term=SAMN02714232
# upper right corner -> send to -> file -> format = full (xml) -> create file
# Parse XML
ncbi_parse(meta = "biosample_result.xml", db = "biosample", format = "xml")#' 
}
}
