% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_parse_assembly_xml.R
\name{ncbi_parse_assembly_xml}
\alias{ncbi_parse_assembly_xml}
\title{Parse NCBI assembly metadata}
\usage{
ncbi_parse_assembly_xml(file)
}
\arguments{
\item{file}{character; path to an xml file.}
}
\value{
a tibble.
}
\description{
This function can be used to parse an xml file from the NCBI assembly
database into a tibble.
}
\examples{
dontrun{
# search for Acinetobacter baumannii within the NCBI Assembly database
# https://www.ncbi.nlm.nih.gov/assembly/?term=acinetobacter\%20baumannii
# upper right corner -> send to -> file -> format = xml -> create file
# parse the downloaded file
ncbi_parse_assembly_xml("assembly_summary.xml")
}
}
