% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_parse_biosample_txt.R
\name{ncbi_parse_biosample_txt}
\alias{ncbi_parse_biosample_txt}
\title{Parse NCBI assembly metadata}
\usage{
ncbi_parse_biosample_txt(file)
}
\arguments{
\item{file}{character; path to an xml file.}
}
\value{
a tibble.
}
\description{
This function can be used to parse an xml file from the NCBI assembly
database into a tibble.
}
\examples{
dontrun{
# search for Acinetobacter baumannii within the NCBI BioSample database
# https://www.ncbi.nlm.nih.gov/biosample/?term=acinetobacter+baumannii
# upper right corner -> send to -> file -> format = full (text) -> create file
# parse the downloaded file
ncbi_parse_biosample_txt("biosample_summary.txt")
}
}
