% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_parse_biosample_xml.R
\name{ncbi_parse_biosample_xml}
\alias{ncbi_parse_biosample_xml}
\title{Parse NCBI BioSample metadata}
\usage{
ncbi_parse_biosample_xml(biosample_xml, verbose = getOption("verbose"))
}
\arguments{
\item{biosample_xml}{character; either a character vector containing an xml
that was retrieved through \code{rentrez::entrez_fetch()} or a path to an xml
file that was downloaded from NCBI BioSample.}

\item{verbose}{logical; Should verbose messages be printed to console?}
}
\description{
BioSample metadata from NCBI can be retrieved in multiple file formats. This
function parses metadata retrieved in XML format.
}
