% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_post_uid.R
\name{ncbi_post_uid}
\alias{ncbi_post_uid}
\title{Post a vector of NCBI UIDs}
\usage{
ncbi_post_uid(uid, db, batch_size = 100, verbose = getOption("verbose"))
}
\arguments{
\item{uid}{numeric; a vector of NCBI UIDs.}

\item{db}{character; the database the UIDs belong to. For options see
`ncbi_dbs()`.}

\item{batch_size}{integer; the number of search terms to query at once. If
the number of search terms is larger than \code{batch_size}, the search terms
are split into batches and queried separately.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
An object of class \code{"ncbi_uid"} which is a list with three
elements:
\itemize{
 \item \code{uid}: a vector of UIDs.
 \item \code{db}: the database the UIDs belong to.
 \item \code{web_history}: a tibble of web histories.
}
}
\description{
This function posts a vector of NCBI UIDs to the NCBI webservice. The purpose
of this function is to generate an ncbi_uid object with web history which can
then be used in downstream functions.
}
\details{
The default value for \code{batch_size} should work in most cases.
However, if the vector of UIDs is very long, the function may fail with an
error message. In this case, try reducing the \code{batch_size} value.
}
\note{
In webseq NCBI UIDs are acquired using `ncbi_get_uid()` or 
`ncbi_link_uid()` and are used for acquiring linked UIDs in other NCBI 
databases or for retrieving the data itself. In some cases data retrieval is
faster when using web history. However, you may have UIDs without a web
history or your web history may expire. In this case you can use 
`ncbi_post_uid()` to generate a new ncbi_uid object with web history.
}
\examples{
ncbi_post_uid(4505768, db = "biosample")
}
