% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_recover_id.R
\name{ncbi_recover_id}
\alias{ncbi_recover_id}
\title{Recover NCBI IDs from NCBI UIDs}
\usage{
ncbi_recover_id(
  query,
  db = NULL,
  batch_size = 100,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{either an object of class \code{ncbi_uid} or an integer vector 
of UIDs. See Details for more information.}

\item{db}{character; the database to search in. For options see
\code{ncbi_dbs()}.}

\item{batch_size}{integer; the number of items to query at once. If query 
length is larger than `batch_size`, the query will be split into batches.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
A vector of matching IDs.
}
\description{
Many functions that interact with NCBI return UIDs. This function converts
the UIDs to NCBI IDs.
}
\details{
If query is an `ncbi_uid` object, the `db` argument is optional. If 
`db` is not specified, the function will retrieve it from the query object. 
However, if it is specified, it must be identical to the `db` attribute of 
the query.
}
\examples{
\dontrun{
uid <- ncbi_get_uid("GCF_000002435.2", db = "assembly")
ncbi_recover_id(uid)
}
}
