% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_header.R
\name{parse_gb_header}
\alias{parse_gb_header}
\title{Parse headers from GenBank files}
\usage{
parse_gb_header(
  file,
  dir = getwd(),
  outfile = "./cache/annotation_headers.rda",
  errorfile = "./cache/annotation_headers_parse_error.rda",
  batch_size = 10,
  verbose = getOption("verbose")
)
}
\arguments{
\item{file}{character;}

\item{dir}{character;}

\item{outfile}{character;}

\item{errorfile}{character;}

\item{batch_size}{integer;}

\item{verbose}{logical;}
}
\description{
Parse headers from GenBank files
}
