% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mlst.R
\name{parse_mlst}
\alias{parse_mlst}
\title{Parse Multi-Locus Sequence Typing output files}
\usage{
parse_mlst(dirpath)
}
\arguments{
\item{dirpath}{character; path to the directory which stores mlst output in
.json format.}
}
\value{
a data.frame
}
\description{
This function will parse the output of MLST predictions into a data.frame
}
