% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_report.R
\name{parse_report}
\alias{parse_report}
\title{Parse assembly reports}
\usage{
parse_report(file)
}
\arguments{
\item{file}{character; the file path to the assembly report.}
}
\value{
The function returns an object of classes \code{arpt} and
\code{list}. The unique class is required for compatibility with subsequent
functions in the pipeline. Otherwise data from the returned object can be
extracted through general list operations.
}
\description{
This function can be used to parse a downloaded assembly report.
}
\note{
This is the fourth step within the pipeline for downloading GenBank
files.
}
\examples{
\dontrun{
phages <- get_genomeid("Autographiviridae", db = "assembly")
report_url <- get_report_url(phages$ids[1])
download_report(report_url)
filename <- dir(paste0(tempdir(), "/assembly_reports"))
filepath <- paste0(tempdir(), "/assembly_reports/", filename)
parse_report(filepath)
}
}
\seealso{
\code{get_genomeid},
\code{get_report_url()},
\code{download_report()},
\code{extract_accn()},
\code{download_gb()}
}
