% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mlst.R
\name{predict_mlst}
\alias{predict_mlst}
\title{Predict Multi-Locus Sequence Type}
\usage{
predict_mlst(config.file, verbose = getOption("verbose"))
}
\arguments{
\item{config.file}{character; path to a local yaml file that stores the job
configurations.}

\item{print}{boolean; verbose messages to console?}
}
\value{
The function will download a set of json files. The output directory
can be defined in the config file.
}
\description{
This function predicts the multi-locus sequence type of a set of sequences
}
\note{
This function depends on \code{singularity} and requires a docker image.
}
\references{
https://sylabs.io/guides/master/user-guide/index.html
}
\seealso{
pull_docker, parse_mlst
}
