% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{dv_download}
\alias{dv_download}
\title{Download files from a Dataset on a Dataverse server}
\usage{
dv_download(files, conn, dirpath = NULL, verbose = getOption("verbose"))
}
\arguments{
\item{files}{a tibble of files from a `dv_connection` object, typically 
obtained from [dv_list_files()] plus optional filtering. Must contain columns
called `filename`, `id` and `md5`.}

\item{conn}{a `dv_connection` object from [dv_connect()].}

\item{dirpath}{character, path to the directory where files will be saved.
Defaults to the current working directory.}

\item{verbose}{logical, should verbose messages be printed to the console?}
}
\value{
A character vector of file paths for successfully downloaded files,
or `NA` for files that failed to download. Returned invisibly.
}
\description{
Download files from a Dataset on a Dataverse server
}
\examples{
conn <- dv_connect(
  dataset = "doi:10.18710/NCKZD7",
  server  = "https://dataverse.no"
)

conn |> 
  dv_list_files() |> 
  dplyr::filter(grepl("\\\\.fasta$", filename)) |>
  dv_download(conn = conn)
}
\seealso{
[dv_connect()], [dv_list_files()]
}
