% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{dv_list_files}
\alias{dv_list_files}
\title{List files in a Dataset of a Dataverse server}
\usage{
dv_list_files(conn, extension = NULL)
}
\arguments{
\item{conn}{a `dv_connection` object from [dv_connect()].}

\item{extension}{character, optional file extension to filter by 
(`"fasta"`, `"fastq"`, etc.).}
}
\value{
A tibble with information about the files in the Dataset.
}
\description{
List files in a Dataset of a Dataverse server
}
\note{
The returned tibbel is not a simple tibble, it is an object of class 
`dv_files` where the connection object is stored as an attribute of the 
tibble. This allows piping the output of `dv_list_files()` directly into 
`dv_download()`.
}
\examples{
conn <- dv_connect(
  dataset = "doi:10.18710/NCKZD7",
  server  = "https://dataverse.no"
)
conn |> dv_list_files(extension = "fasta")
}
\seealso{
[dv_connect()], [dv_download()]
}
