% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{examples}
\alias{examples}
\title{Examples}
\format{
A list with 6 elements:
\describe{
  \item{assembly}{NCBI Assembly IDs}
  \item{bioproject}{NCBI BioProject IDs}
  \item{biosample}{NCBI BioSample IDs}
  \item{gene}{NCBI Gene IDs}
  \item{protein}{NCBI Protein IDs}
  \item{sra}{NCBI SRA IDs}
}
}
\usage{
examples
}
\description{
This data set contains a list of IDs which can be used to access data from
various data sources. These IDs are used across the package in function
documentations, tests, vignettes.
}
\keyword{datasets}
