% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uid.R
\name{get_uid}
\alias{get_uid}
\title{Collect internal IDs from NCBI databases}
\usage{
get_uid(term, db = "assembly")
}
\arguments{
\item{term}{character; the search term.}

\item{db}{character; the database to search in. For options see
\code{rentrez::entrez_dbs()}}
}
\value{
a character vector of ID-s.
}
\description{
This function replicates the NCBI website's search utility. The search term
is matched against the chosen database and the function returns a character
vector of internal IDs that can be used e.g. to link NCBI entries with
entries in other NCBI databases.
}
\examples{
\dontrun{
get_uid("GCA_003012895.2")
get_uid("Autographiviridae OR Podoviridae", db = "assembly")
}
}
