% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_download_genome.R
\name{ncbi_download_genome}
\alias{ncbi_download_genome}
\title{Download Genomes from NCBI}
\usage{
ncbi_download_genome(
  accession,
  type = "genomic.gbff",
  dirpath = NULL,
  verbose = getOption("verbose")
)
}
\arguments{
\item{accession}{character; a character vector of assembly accessions.}

\item{type}{character; the file extension to download. Valid options are
\code{"assembly_report"}, \code{"assembly_stats"}, \code{"cds"},
\code{"feature_count"}, \code{"feature_table"}, \code{"genomic.fna"},
\code{"genomic.gbff"}, \code{"genomic.gff"}, \code{"genomic.gtf"},
\code{"protein.faa"}, \code{"protein.gpff"}, \code{"translated_cds"}.}

\item{dirpath}{character; the path to the directory where the file should be
downloaded. If \code{NULL}, download file to the working directory.}

\item{verbose}{logical; should verbose messages be printed to console?}
}
\description{
This function directly downloads genome data through the NCBI FTP server.
}
\examples{
\dontrun{
# Download genbank file for GCF_003007635.1.
# The function will access files within this directory:
# ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/003/007/635/
ncbi_download_genome("GCF_003007635.1", type = "genomic.gbff", verbose = TRUE)

# Download multiple files
accessions <- c("GCF_000248195.1", "GCF_000695855.3")
ncbi_download_genome(accessions, type = "genomic.gbff", verbose = TRUE)
}
}
