% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_get_meta.R
\name{ncbi_get_meta}
\alias{ncbi_get_meta}
\title{Get sequence metadata from NCBI}
\usage{
ncbi_get_meta(
  query,
  db = NULL,
  batch_size = 100,
  use_history = TRUE,
  parse = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{either an object of class \code{ncbi_uid} or an integer vector 
of UIDs. See Details for more information.}

\item{db}{character; the database to search in. For options see
\code{ncbi_dbs()}. Not all databases are supported.}

\item{batch_size}{integer; the number of search terms to query at once. If
the number of search terms is larger than \code{batch_size}, the search terms
are split into batches and queried separately.}

\item{use_history}{logical; should the function use web history for faster
API queries?}

\item{parse}{logical; Should the function attempt to parse the output into a
tibble? If unsuccessful, the function will return the unparsed output.}

\item{verbose}{logical; Should verbose messages be printed to console?}
}
\value{
The function returns a list with two elements:
\itemize{
\item \code{meta}: if \code{parse = TRUE} then either a tibble with the 
metadata or if parsing is unsuccessful, the unparsed metadata. If
\code{parse = FALSE} the unparsed metadata.
\item \code{history}: a tibble of web histories.
}
}
\description{
This function retrieves metadata from a given NCBI sequence database.
}
\details{
Some functions in webseq, e.g. \code{ncbi_get_uid()} or
\code{ncbi_link_uid()} return objects of class \code{"ncbi_uid"}. These
objects may be used directly as query input for \code{ncbi_get_meta()}. This
approach is recommended because the internal structure of these objects make
\code{ncbi_get_meta()} queries more robust. Alternatively, you can also
use a character vector of UIDs as query input.

If query is a \code{"ncbi_uid"} object, the \code{db} argument is
optional. If \code{db} is not specified, the function will use the
\code{db} attribute of the \code{"ncbi_uid"} object as \code{db} argument.
However, if it is specified, it must be identical to the \code{db} attribute
of the \code{"ncbi_uid"} object. If query is a character vector, the
\code{db} argument is required.
}
\examples{
\dontrun{
data(examples)
uids <- ncbi_get_uid(examples$biosample, db = "biosample")
meta <- ncbi_get_meta(uids)
}
}
