% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_get_meta.R
\name{ncbi_get_meta}
\alias{ncbi_get_meta}
\title{Get sequence metadata from NCBI}
\usage{
ncbi_get_meta(id, db, batch_size = 250, verbose = getOption("verbose"))
}
\arguments{
\item{id}{integer; an integer vector of database specific NCBI UIDs.}

\item{db}{character; the database to search in. For options see
\code{rentrez::entrez_dbs()}.}

\item{batch_size}{integer; the number of search terms to query at once. If the
number of search terms is larger than \code{batch_size}, the search terms
are split into batches and queried separately.}

\item{verbose}{logical; Should verbose messages be printed to console?}
}
\description{
This function is a wrapper for \code{rentrez::entrez_fetch()} that retrieves
metadata from a given NCBI sequence database. The function currently works
with the following databases: \code{"assembly"}, \code{"biosample"}.
}
\examples{
\dontrun{
data(examples)
uids <- ncbi_get_uid(examples$biosample, db = "biosample")
meta <- ncbi_get_meta(uids$uid, db = "biosample")
}
}
