% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_get_uid.R
\name{ncbi_get_uid}
\alias{ncbi_get_uid}
\title{Get UID-s from NCBI databases}
\usage{
ncbi_get_uid(
  term,
  db,
  batch_size = 100,
  use_history = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{term}{character; one or more search terms.}

\item{db}{character; the database to search in. For options see
\code{ncbi_dbs()}}

\item{batch_size}{integer; the number of search terms to query at once. If
the number of search terms is larger than \code{batch_size}, the search terms
are split into batches and queried separately.}

\item{use_history}{logical; should the function use web history for faster
API queries?}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
The function returns a list with two elements:
\itemize{
 \item \code{uids}: a tibble with the search terms and their corresponding
 UIDs.
 \item \code{history}: if \code{web_history = TRUE}, a character vector with
 the web history IDs of the individual batches, otherwise \code{NULL}.
 }
}
\description{
This function replicates the NCBI website's search utility. It searches one
or more search terms in the chosen database and returns internal NCBI UID-s 
for individual hits. These can be used e.g. to link NCBI entries with entries
in other NCBI databases or to retrieve the data itself.
}
\details{
The default value for \code{batch_size} should work in most cases.
However, if the search terms are very long, the function may fail with an
error message. In this case, try reducing the \code{batch_size} value.
}
\examples{
\dontrun{
ncbi_get_uid("GCA_003012895.2", db = "assembly")
ncbi_get_uid("Autographiviridae OR Podoviridae", db = "biosample")
ncbi_get_uid(c("WP_093980916.1", "WP_181249115.1"), db = "protein")
}
}
