% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_get_uid.R
\name{ncbi_get_uid}
\alias{ncbi_get_uid}
\title{Get UID-s from NCBI databases}
\usage{
ncbi_get_uid(
  term,
  db,
  batch_size = 100,
  use_history = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{term}{character; one or more search terms.}

\item{db}{character; the database to search in. For options see
\code{ncbi_dbs()}}

\item{batch_size}{integer; the number of search terms to query at once. If
the number of search terms is larger than \code{batch_size}, the search terms
are split into batches and queried separately. Not used when using web
history.}

\item{use_history}{logical; should the function use web history for faster
API queries?}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
An object of class \code{"ncbi_uid"} which is a list with three
elements:
\itemize{
 \item \code{uids}: a vector of UIDs.
 \item \code{db}: the database used for the query.
 \item \code{web_history}: if \code{web_history = TRUE}, a tibble which
 contains information about the web history, otherwise \code{NULL}.
 }
}
\description{
This function replicates the NCBI website's search utility. It searches one
or more search terms in the chosen database and returns internal NCBI UID-s 
for the hits. These can be used e.g. to link NCBI entries with entries in 
other NCBI databases or to retrieve the data itself.
}
\details{
The default value for \code{batch_size} should work in most cases.
However, if the search terms are very long, the function may fail with an
error message. In this case, try reducing the \code{batch_size} value.
}
\examples{
\dontrun{
ncbi_get_uid("GCA_003012895.2", db = "assembly")
ncbi_get_uid("Autographiviridae OR Podoviridae", db = "biosample")
ncbi_get_uid(c("WP_093980916.1", "WP_181249115.1"), db = "protein")
}
}
