% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_link_uid.R
\name{ncbi_link_uid}
\alias{ncbi_link_uid}
\title{Link UID-s from one NCBI database to another}
\usage{
ncbi_link_uid(
  query,
  from = NULL,
  to,
  batch_size = 100,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{either an object of class \code{ncbi_uid} or an integer vector 
of UIDs. See Details for more information.}

\item{from}{character; the database the queried UIDs come from.
\code{ncbi_dbs()} lists all available options.}

\item{to}{character; the database in which the function should look for links.
\code{ncbi_dbs()} lists all available options. See Details for more
information.}

\item{batch_size}{integer; the number of search terms to query at once. If
the number of search terms is larger than \code{batch_size}, the search terms
are split into batches and queried separately. Not used when using web
history.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
A tibble with two columns. The first column contains UIDs in the 
`from` database, the second column contains linked UIDs in the `to` database.
}
\description{
Each entry in an NCBI database has its unique internal id. Entries in
different databases may be linked. For example, entries in the NCBI Assembly
database may be linked with entries in the NCBI BioSample database. This
function attempts to link uids from one database to another.
}
\details{
The function `ncbi_get_uid()` returns an object of class `ncbi_uid`. 
This object may be used directly as query for `ncbi_link_uid()`. If query is 
an `ncbi_uid` object, the `from` argument is optional. If `from` is not 
specified, the function will retrieve it from the query object. However, if 
it is specified, it must be identical to the `db` attribute of the query.
}
\examples{
ncbi_link_uid(5197591, "assembly", "biosample")
ncbi_link_uid(c(1226742659, 1883410844), "protein", "nuccore")
}
