% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_link_uid.R
\name{ncbi_link_uid}
\alias{ncbi_link_uid}
\title{Link UID-s from one NCBI database to another}
\usage{
ncbi_link_uid(
  query,
  from = NULL,
  to,
  batch_size = 100,
  use_history = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{either an object of class \code{ncbi_uid} or an integer vector 
of UIDs. See Details for more information.}

\item{from}{character; the database the queried UIDs come from.
\code{ncbi_dbs()} lists all available options.}

\item{to}{character; the database in which the function should look for links.
\code{ncbi_dbs()} lists all available options. See Details for more
information.}

\item{batch_size}{integer; the number of search terms to query at once. If
the number of search terms is larger than \code{batch_size}, the search terms
are split into batches and queried separately. Not used when using web
history.}

\item{use_history}{logical; should the function use web history for faster
API queries?}

\item{verbose}{logical; should verbose messages be printed to the console?
\code{ncbi_dbs()} lists all available options.}
}
\value{
An object of class \code{"ncbi_uid"} which is a list with three
elements:
\itemize{
 \item \code{uid}: a vector of UIDs.
 \item \code{db}: the database used for the query.
 \item \code{web_history}: a tibble of web histories.
 }
}
\description{
Each entry in an NCBI database has its unique internal id. Entries in
different databases may be linked. For example, entries in the NCBI Assembly
database may be linked with entries in the NCBI BioSample database. This
function attempts to link uids from one database to another.
}
\details{
Some functions in webseq, e.g. \code{ncbi_get_uid()} or
\code{ncbi_link_uid()} return objects of class \code{"ncbi_uid"}. These
objects may be used directly as query input for \code{ncbi_link_uid()}. This
approach is recommended because the internal structure of these objects make
\code{ncbi_link_uid()} queries more robust. Alternatively, you can also
use a character vector of UIDs as query input.

If query is a \code{"ncbi_uid"} object, the \code{from} argument is
optional. If \code{from} is not specified, the function will use the
\code{db} attribute of the \code{"ncbi_uid"} object as \code{from} argument.
However, if it is specified, it must be identical to the \code{db} attribute
of the \code{"ncbi_uid"} object. If query is a character vector, the
\code{from} argument is required.
}
\examples{
\dontrun{
ncbi_link_uid("4253631", "assembly", "biosample")
ncbi_link_uid(c("1226742659", "1883410844"), "protein", "nuccore")
}
}
