% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_parse.R
\name{ncbi_parse}
\alias{ncbi_parse}
\title{Parse NCBI sequence metadata}
\usage{
ncbi_parse(meta, db, format = "xml", verbose = getOption("verbose"))
}
\arguments{
\item{meta}{character; either a character vector containing a data set that
was retrieved through \code{rentrez::entrez_fetch()} or a path to an file
that was downloaded from NCBI.}

\item{db}{character; the NCBI database from which the data was retrieved.
Currently only \code{"biosample"} is supported.}

\item{format}{character; the format of the data set. Currently only
\code{"xml"} is supported.}

\item{verbose}{logical; Should verbose messages be printed to console?}
}
\description{
This function can be used to parse various non-sequence data sets from NCBI
into a tibble. These data sets usually accompany the biological sequences and
contain additional information e.g. identifiers, information about the
sample, the sequencing platform, etc. The function currently supports parsing
NCBI BioSample data from XML format.
}
\examples{
\dontrun{
data(examples)
#' 
# NCBI Assembly, download XML file from NCBI and parse

# Manually download the XML file
# https://www.ncbi.nlm.nih.gov/assembly/GCF_000299415.1
# upper right corner -> send to -> file -> format = xml -> create file
# Parse XML
ncbi_parse(meta = "assembly_summary.xml", db = "assembly", format = "xml")

# NCBI BioSample, fully programmatic access

# Get internal BioSample UID for BioSample ID
biosample_uid <- ncbi_get_uid(examples$biosample, db = "biosample")
# Get metadata in XML format
meta <- ncbi_get_meta(uids$uid, db = "biosample")
# Parse XML
ncbi_parse(meta = meta_xml, db = "biosample", format = "xml")

# NCBI BioSample, download XML file from NCBI and parse

# Manually download the XML file
# https://www.ncbi.nlm.nih.gov/biosample/?term=SAMN02714232
# upper right corner -> send to -> file -> format = full (xml) -> create file
# Parse XML
ncbi_parse(meta = "biosample_result.xml", db = "biosample", format = "xml")#' 
}
}
