% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_parse_biosample_txt.R
\name{ncbi_parse_biosample_txt}
\alias{ncbi_parse_biosample_txt}
\title{Parse NCBI BioSample metadata}
\usage{
ncbi_parse_biosample_txt(
  file,
  resolve_na = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{file}{character; path to a txt file.}

\item{resolve_na}{logical; replace strings that match NA terms with NA.}

\item{verbose}{logical; should verbose output be printed to console?}
}
\value{
a tibble.
}
\description{
This function parses a txt file from the NCBI BioSample database.
}
\examples{
dontrun{
# search for Acinetobacter baumannii within the NCBI BioSample database
# https://www.ncbi.nlm.nih.gov/biosample/?term=acinetobacter+baumannii
# upper right corner -> send to -> file -> format = full (text) -> create file
# parse the downloaded file
ncbi_parse_biosample_txt("biosample_summary.txt")
}
}
